Replaced StaticMap with Snapshot and SnapshotOptions
Overhauled the code to ever-so-vaguely resemble MapKit’s MKMapSnapshotter API. A SnapshotOptions object and access token are passed into the Snapshot initializer, which builds a URL request and can fetch the image either synchronously or asynchronously. Instead of setting an “autoFitFeatures” parameter, auto-fitting is enabled when the center coordinate is absent – Null Island is no longer the default center point.

If no access token is passed into the Snapshot initializer, it is read from the Info.plist, just like the Mapbox iOS SDK does.

Increased usage of NSURL and NSURLComponents to ensure correct URL handling. Added an optional NSError to the completion handler, in case things go awry.